requirements.txt
imageio==2.33.0
jieba==0.42.1
numpy==1.26.4
pandas==2.2.1
PyMySQL==1.1.0
wordcloud==1.9.3
schedule==1.2.1
loguru==0.7.2
cryptography==42.0.5
Pillow==9.5.0
